update StealthMine.java - fixed swapping issues, should now work fine - updated the settings
added PvpInfo.java
- added "ran away" method (detects if a player has ran away)
- added "logout" (detects if a player logged out during combat) STILL W.I.P

+++ src/main/java/org/Snail/Plus/utils/SwapUtils.java
@@ -1,21 +1,12 @@
 package org.Snail.Plus.utils;
 
-import it.unimi.dsi.fastutil.ints.Int2ObjectArrayMap;
-import meteordevelopment.meteorclient.mixininterface.IClientPlayerInteractionManager;
 import meteordevelopment.meteorclient.utils.player.InvUtils;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.item.ItemStack;
-import net.minecraft.network.packet.c2s.play.ClickSlotC2SPacket;
-import net.minecraft.screen.ScreenHandler;
-import net.minecraft.screen.slot.SlotActionType;
-
-import static meteordevelopment.meteorclient.MeteorClient.mc;
 
 
 public class SwapUtils {
     private static long lastTimeoutCheck = 0;
 
-    public static void SilentSwap(int ItemSlot, float Delay) {
+    public static void SilentSwap(int ItemSlot, Double Delay) {
 
         InvUtils.swap(ItemSlot, true);
